k380: set output pins to be disconnected or high
The K380 doesn't have any diodes in the keyboard matrix, so when a
keyboard scan is done and two keys on the same column are pressed, we
would normally short two output / row pins. One would be high and the
other low.

This can be prevented by setting the pins to be GPIO_OPEN_SOURCE as now
the pin will have high impedance / behave disconnected if driven low /
inactive.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>