kscan: Limit maximum amount of key presses to 3
When using a keyboard without diodes in the key matrix, ghosting can
become a serious issue. Quite often three keys held at the same time are
needed to produce a ghosted fourth key press when using a row, col based
keyboard matrix. In these cases it makes sense to allow only three key
presses at the same time, ignoring all of them if a fourth is pressed.

The changes made here are a simple work-of-proof how this "limit the
amount of key presses" could be done.

Note: There's no configuration options or similar to change this
behavior right now.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>