Quick Fix: Mod Morph: iOS better modifier announce
See https://github.com/zmkfirmware/zmk/issues/1718

But the basic idea is that iPad (and presumably iOS too) will only
partially handle mod morph (and like) correctly if the modifiers are
announced directly with the new key press.

Forcing zmk to send a keyboard report with the new modifiers before
sending the triggered key resolves this issue and makes usage of mod
morph consistently stable.

This issue probably also happens with all behaviors, so this is only a
very quick and dirty fix for it.

Signed-off-by: Sophie Friedrich <dev@flowerpot.me>