[WIP] feat: Add TyK380v2 keyboard definition
The TyK380v2 is some custom firmware for the Logitech K380s keyboard.
Currently the keyboard matrix, leds and co have yet to be reverse
engineered, but this board builds (currently untested whether it runs).

The current keyboard matrix definition is one exemplary way in which we
could reverse engineer the keyboard matrix. By setting a single row and
then looking which key will get activated by the columns.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>