feat: Enable persistent settings using flash (8kb) and code part
Due to the NVE driver requiring two full flash sectors in order to
maintain a consistent state, the storage partition needed to be
increased to 8kb.

Also enabled (and enforced) code partition in order to make sure that we
don't flash code that breaks instantly as it gets overwritten by the
settings process.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>