feat: Enable RAM and ROM optimizations
The application would previously not fit into the limited resources of
the nRF52832_QFAB platform. With these settings there is still enough
space left for ZMK to work properly.

The biggest changes have been decreasing the heap as it is not used,
changing the simultaneous BT connections to three (as each requires a
large chunk of the RAM) and optimizing for ROM size instead of debug.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>